+++ docs/html/search/all_c.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['segmentdriver',['SegmentDriver',['../classace__segment_1_1SegmentDriver.html',1,'ace_segment::SegmentDriver'],['../classace__segment_1_1SegmentDriver.html#ac5c0131ae8329e297db5314594108451',1,'ace_segment::SegmentDriver::SegmentDriver()']]],
+  ['segmentdriver',['SegmentDriver',['../classace__segment_1_1SegmentDriver.html',1,'ace_segment::SegmentDriver'],['../classace__segment_1_1SegmentDriver.html#ab34b3e012f551568b94e6d573eff4659',1,'ace_segment::SegmentDriver::SegmentDriver()']]],
   ['segmentpatterntype',['SegmentPatternType',['../classace__segment_1_1Driver.html#a851393f9d1f557871e259b295e733920',1,'ace_segment::Driver']]],
   ['setanodeongroup',['setAnodeOnGroup',['../classace__segment_1_1LedMatrix.html#a562150f95e1397ddc8bafe54ac5f2a71',1,'ace_segment::LedMatrix']]],
   ['setblinkfastduration',['setBlinkFastDuration',['../classace__segment_1_1RendererBuilder.html#a164c3a5bbfe35f59b310abd8d3704880',1,'ace_segment::RendererBuilder']]],
@@ -8,7 +8,6 @@ var searchData=
   ['setbrightness',['setBrightness',['../classace__segment_1_1Driver.html#aba1b47db2304b6244f4759a959125a9f',1,'ace_segment::Driver']]],
   ['setcathodeongroup',['setCathodeOnGroup',['../classace__segment_1_1LedMatrix.html#a0dc49954ce07d860d297bbdd1d4aacb7',1,'ace_segment::LedMatrix']]],
   ['setframespersecond',['setFramesPerSecond',['../classace__segment_1_1RendererBuilder.html#af0c9855f22aab430b469c16779b8253b',1,'ace_segment::RendererBuilder']]],
-  ['setnumsubfields',['setNumSubFields',['../classace__segment_1_1DriverBuilder.html#a7f0ad00b3791bb28ee1662b1a167cb23',1,'ace_segment::DriverBuilder']]],
   ['setpattern',['setPattern',['../classace__segment_1_1Driver.html#a84f21c38b72a4c576a44e2d370356520',1,'ace_segment::Driver']]],
   ['setpulsefastduration',['setPulseFastDuration',['../classace__segment_1_1RendererBuilder.html#a0e6113e40832ebd4eb5572e00712ea9a',1,'ace_segment::RendererBuilder']]],
   ['setpulseslowduration',['setPulseSlowDuration',['../classace__segment_1_1RendererBuilder.html#a5cc5bb2635f8a7837f34a3a12aa46ded',1,'ace_segment::RendererBuilder']]],