+++ docs/html/search/classes_2.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['hardware',['Hardware',['../classace__segment_1_1Hardware.html',1,'ace_segment']]]
+  ['hardware',['Hardware',['../classace__segment_1_1Hardware.html',1,'ace_segment']]],
+  ['hexwriter',['HexWriter',['../classace__segment_1_1HexWriter.html',1,'ace_segment']]]
 ];