+++ docs/html/search/functions_6.js
@@ -2,5 +2,5 @@ var searchData=
 [
   ['micros',['micros',['../classace__segment_1_1Hardware.html#a577bbefb50970431da768e99e8765798',1,'ace_segment::Hardware']]],
   ['millis',['millis',['../classace__segment_1_1Hardware.html#a967adfe2fb7acb7eecca011e34a4a690',1,'ace_segment::Hardware']]],
-  ['modulatingdigitdriver',['ModulatingDigitDriver',['../classace__segment_1_1ModulatingDigitDriver.html#ac2ebc2b78610a703de8c8481f773fecc',1,'ace_segment::ModulatingDigitDriver']]]
+  ['modulatingdigitdriver',['ModulatingDigitDriver',['../classace__segment_1_1ModulatingDigitDriver.html#acb7ea555663d194616f47d9605956c8d',1,'ace_segment::ModulatingDigitDriver']]]
 ];