+++ docs/html/search/functions_b.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['updateframe',['updateFrame',['../classace__segment_1_1Renderer.html#aa0151565ff7b9f4fbd41f799b6e85186',1,'ace_segment::Renderer']]]
+  ['updateframe',['updateFrame',['../classace__segment_1_1Renderer.html#aa0151565ff7b9f4fbd41f799b6e85186',1,'ace_segment::Renderer']]],
+  ['usemodulatingdriver',['useModulatingDriver',['../classace__segment_1_1DriverBuilder.html#a8303f5066da98e26189047d5623ce64b',1,'ace_segment::DriverBuilder']]]
 ];