+++ src/ace_segment/ModulatingDigitDriver.h
@@ -37,8 +37,8 @@ class ModulatingDigitDriver: public DigitDriver {
     /** Constructor. */
     explicit ModulatingDigitDriver(LedMatrix* ledMatrix,
             DimmingDigit* dimmingDigits, uint8_t numDigits,
-            uint8_t numSubFields):
-        DigitDriver(ledMatrix, dimmingDigits, numDigits),
+            uint8_t numSubFields, bool ownsLedMatrix = false):
+        DigitDriver(ledMatrix, dimmingDigits, numDigits, ownsLedMatrix),
         mNumSubFields(numSubFields)
     {}
 
@@ -56,14 +56,15 @@ class ModulatingDigitDriver: public DigitDriver {
 
     virtual void displayCurrentField() override;
 
+  protected:
+    uint8_t const mNumSubFields;
+    uint8_t mCurrentSubField;
+    uint8_t mCurrentSubFieldMax;
+
   private:
     // disable copy-constructor and assignment operator
     ModulatingDigitDriver(const ModulatingDigitDriver&) = delete;
     ModulatingDigitDriver& operator=(const ModulatingDigitDriver&) = delete;
-
-    uint8_t const mNumSubFields;
-    uint8_t mCurrentSubField;
-    uint8_t mCurrentSubFieldMax;
 };
 
 }