+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+* 0.2.0 (2018-04-12)
+    * Support `digitalWriteFast()` through `fast_driver.py` code generator.
+    * Gather flash and static memory consumption numbers into README.md.
+    * Add `HexWriter` class, useful when `CharWriter` is overkill.
+    * Support resource creation without `DriverBuilder` by storing flag to
+      indicate memory ownership.
 * 0.1.0 (2018-04-10)
     * Support 74HC595 serial-to-parallel converter chip (fixes #3).
     * Add unit tests for all major layers.